LPS-105380 Move classes /internal/configuration/admin/display/ to /internal/portal/settings/configuration/admin/display/
