+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/ContentSharingEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
 
@@ -23,22 +23,22 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
 public class
-	EmailVerificationNotificationEditCompanyPortalSettingsConfigurationScreenContributor
+	ContentSharingEditCompanyPortalSettingsConfigurationScreenContributor
 		extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "email";
+		return "sharing";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/email.notifications/email_verification_notification.jsp";
+		return "/content_sharing.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "email-verification-notification";
+		return "content-sharing";
 	}
 
 }
