+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/EmailSenderEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,18 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
-import com.liferay.map.constants.MapProviderWebKeys;
-import com.liferay.map.util.MapProviderHelperUtil;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
-
-import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -31,42 +22,22 @@ import org.osgi.service.component.annotations.Component;
  * @author Drew Brokke
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class MapsEditCompanyPortalSettingsConfigurationScreenContributor
+public class EmailSenderEditCompanyPortalSettingsConfigurationScreenContributor
 	extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "third-party";
+		return "email";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/maps.jsp";
+		return "/email.notifications/email_sender.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "third-party-maps";
-	}
-
-	@Override
-	public String getName(Locale locale) {
-		return "maps";
-	}
-
-	@Override
-	public void setAttributes(
-		HttpServletRequest httpServletRequest,
-		HttpServletResponse httpServletResponse) {
-
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		httpServletRequest.setAttribute(
-			MapProviderWebKeys.MAP_PROVIDER_KEY,
-			MapProviderHelperUtil.getMapProviderKey(
-				themeDisplay.getCompanyId()));
+		return "email-sender";
 	}
 
 }
