+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/LanguageEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,48 +12,32 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
-import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
 
-import java.util.Locale;
-
-import javax.servlet.ServletContext;
-
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Drew Brokke
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class LegacyEditCompanyPortalSettingsConfigurationScreen
+public class LanguageEditCompanyPortalSettingsConfigurationScreenContributor
 	extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "instance-configuration";
+		return "localization";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/general.jsp";
+		return "/language.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "general";
+		return "language";
 	}
-
-	@Override
-	public String getName(Locale locale) {
-		return LanguageUtil.get(locale, "general");
-	}
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.portal.settings.web)"
-	)
-	private ServletContext _servletContext;
 
 }
