+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/MapsEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,9 +12,18 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
+import com.liferay.map.constants.MapProviderWebKeys;
+import com.liferay.map.util.MapProviderHelperUtil;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
+
+import java.util.Locale;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -22,22 +31,42 @@ import org.osgi.service.component.annotations.Component;
  * @author Drew Brokke
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class AnalyticsEditCompanyPortalSettingsConfigurationScreenContributor
+public class MapsEditCompanyPortalSettingsConfigurationScreenContributor
 	extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "analytics";
+		return "third-party";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/analytics.jsp";
+		return "/maps.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "analytics";
+		return "third-party-maps";
+	}
+
+	@Override
+	public String getName(Locale locale) {
+		return "maps";
+	}
+
+	@Override
+	public void setAttributes(
+		HttpServletRequest httpServletRequest,
+		HttpServletResponse httpServletResponse) {
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		httpServletRequest.setAttribute(
+			MapProviderWebKeys.MAP_PROVIDER_KEY,
+			MapProviderHelperUtil.getMapProviderKey(
+				themeDisplay.getCompanyId()));
 	}
 
 }
