+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/RatingsEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,9 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
+import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionUtil;
+import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionValues;
+
+import java.util.Locale;
+import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -22,23 +27,41 @@ import org.osgi.service.component.annotations.Component;
  * @author Drew Brokke
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class
-	DefaultUserAssociationsEditCompanyPortalSettingsConfigurationScreenContributor
-		extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
+public class RatingsEditCompanyPortalSettingsConfigurationScreenContributor
+	extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "users";
+		return "community-tools";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/users/default_user_associations.jsp";
+		return "/ratings.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "default-user-associations";
+		return "social-ratings";
+	}
+
+	@Override
+	public String getName(Locale locale) {
+		return "ratings";
+	}
+
+	@Override
+	public boolean isVisible() {
+		Map<String, PortletRatingsDefinitionValues>
+			portletRatingsDefinitionValuesMap =
+				PortletRatingsDefinitionUtil.
+					getPortletRatingsDefinitionValuesMap();
+
+		if (portletRatingsDefinitionValuesMap.isEmpty()) {
+			return false;
+		}
+
+		return true;
 	}
 
 }
