+++ modules/apps/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/internal/portal/settings/configuration/admin/display/TimeZoneEditCompanyPortalSettingsConfigurationScreenContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.configuration.admin.display;
+package com.liferay.portal.settings.web.internal.portal.settings.configuration.admin.display;
 
 import com.liferay.portal.settings.configuration.admin.display.PortalSettingsConfigurationScreenContributor;
 
@@ -22,23 +22,22 @@ import org.osgi.service.component.annotations.Component;
  * @author Drew Brokke
  */
 @Component(service = PortalSettingsConfigurationScreenContributor.class)
-public class
-	MailHostNamesEditCompanyPortalSettingsConfigurationScreenContributor
-		extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
+public class TimeZoneEditCompanyPortalSettingsConfigurationScreenContributor
+	extends BaseEditCompanyPortalSettingsConfigurationScreenContributor {
 
 	@Override
 	public String getCategoryKey() {
-		return "email";
+		return "localization";
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/email.notifications/mail_host_names.jsp";
+		return "/time_zone.jsp";
 	}
 
 	@Override
 	public String getKey() {
-		return "mail-host-names";
+		return "time-zone";
 	}
 
 }
